Initial modifications for nvcc instead of nvc++

+++ src_clean/mc_swap_moves.h
@@ -494,7 +494,7 @@ static inline MoveEnergy IdentitySwapMove(Components& SystemComponents, Simulati
   //Take care of the frist bead location HERE//
   copy_firstbead_to_new<<<1,1>>>(Sims.New, Sims.d_a, OLDComponent, OLDMolInComponent * SystemComponents.Moleculesize[OLDComponent]);
   //WRITE THE component and molecule ID THAT IS GOING TO BE IGNORED DURING THE ENERGY CALC//
-  Sims.ExcludeList[0] = {OLDComponent, OLDMolInComponent};
+  Sims.ExcludeList[0] = {(int)OLDComponent, (int)OLDMolInComponent};
   double2 newScale  = SystemComponents.Lambda[NEWComponent].SET_SCALE(1.0); //Zhao's note: not used in reinsertion, just set to 1.0//
   double Rosenbluth=Widom_Move_FirstBead_PARTIAL(SystemComponents, Sims, FF, Random, Widom, NEWMolInComponent, NEWComponent, CBMCType, StoredR, &SelectedTrial, &SuccessConstruction, &energy, newScale);
